Small fixes and lint update
